перенес getAlikeMovies в селекторы, убрал магическое число
